requirements.txt
streamlit
opencv-python
pytesseract
pillow
pandas
matplotlib
python-dotenv
groq
